Significant updates to PCEAS symbol handling.
Rework handling of bank information in symbols, so that both the
"rombank" and "mprbank" values are stored.

Consolidate all of the MPR-vs-ROM bank calculations into a small
set of routines in symbol.c that are called whenever needed, and
so simplify updates in the future.

Add ".opt f+" to allow forward-references in .EQU and .DS, which
causes a 3rd pass to resolve the symbols. Default this option to
enabled for now because it is friendlier for new programmers.

Add ".3pass" to force a 3rd pass even when PCEAS doesn't believe
that one is needed. This can be used to set the size and address
of stuff at the end of the 1st pass when data sizes are known.

Remove ".overlay", because it is not needed anymore.

Change prefix-keyword expression handling to have 3 underscores
in PCEAS in order to use 2 underscores to access them in C code.