Major rework of the PCEAS code that strips out unused procedures.
The stripping decision has been moved out of proc_reloc(), which is
run just before the final pass, and into a new proc_strip() routine
that is run at the end of the first pass.

Stripped procedures are not printed in the listing file anymore, but
are still assembled into the STRIPPED_BANK which is not output, that
is what I refer to as them being "cloaked".

When PCEAS has decided to run 3 passes, then stripped procedures can
instead be "skipped", i.e. not assembled at all. This allows removal
of unused .procs within a .procgroup, which could not be done before.

Skipping procedures can alter the order and location of symbols that
are defined, and so it requires an extra pass, and it isn't possible
if the start and end of the .proc are wrapped in complex sections of
".if" code.