Small fixes and cleanup to xgetopt source.
All original xopt* and xget* labels have changed to opt* and get* so that this
is a drop-in replacement for the GNU/BSD functions on Windows.

Error messages have been changed to make them more consistent with MSYS2.

In getopt_long_only(), ambiguous single letter options now try short options.