Alter PCEAS to combine multiple source files on the command line into a single build.
The extra source files on the command line are automatically included at the end of
the main (first) source file, and the section is changed to .CODE before each one.

These files are included before "kickc-final" or "sdcc-final".

While this could be used to include asm library code, it was added so that seperately
compiled C source files could be combined into a single project output, with PCEAS as
the "linker", but using assembly language source instead of object files.

This also includes a bunch of cleanup and error-checking of the command line options.