Initial changes from HuC to HuCC.
Revert as much as possible of the change from "int" to "intptr_t" that
was done in a hurry a few years ago to allow a 64-bit build of HuC.

This also *begins* the process of cleaning up some stuff in the SmallC
codebase that should have been "bool" or something other than an "int".

Removed the remains of Uli's interrupt-handling additions.

Begin renaming some of the macros as part of figuring out how to fix
the code-generation for conditionals.