Small tweaks to the international fonts in examples/asm/elmer/fonts/
Unify '!' character in the fonts.

Add a Japanese iso646 font (which just changes the \ to a Yen symbol).

Add an ETS WST (World Standard Teletext) font for Portuguese and Spanish
accented lower-case because the Portuguese iso646 font only has accented
upper-case characters which won't work with the font's 8-pixel height.