Change HuCC function prolog and epilog code generation, and the parameter-passing method when calling functions.
The calling function now puts all parameters on the stack instead
of expecting the called function to put the final parameter on it.

If a C function returns a value, it stores the lo-byte in ZP, and
the caller retrieves it from ZP.

Change I_ADDMI to I_MODSP to better reflect what it is used for.

Change indirect function-pointer calls from I_CALLS to I_CALLP to
avoid putting the function-pointer on the C stack and thus having
to remove it from the C stack afterwards.