Remove I_LDBY and I_LDYB i-codes that seem to have been a mostly unsuccessful attempt to optimize access to a byte-array.
Remove X_LDB_P and X_LDUB_P i-codes which only seem to exist in order to use
zero-page access to the __ptr variable. PCEAS automatically handles that now.

Change the optimizer so that the X_LDB_S optimization is done in the routine
that handles the X_LDW_S optimization, instead of relying on X_LDB_P.