Beautify HuCC's code generation in order to make it easier to read.
Reorder the i-codes in defs.h to group similar functions together so
that you can see what kind of alternatives exist.

Reorder both the macros in hucc-codegen.asm, and the output routines
code.c, to match the order of i-codes in defs.h to make finding them
a bit simpler.

Remove some i-codes that were never actually generated.

Clean up optimize.c so that all of the peephole tests have a similar
structure and comment layout.

Change HuCC's code generation of "switch()" and "case" to output new
__switch, __case and __endcase macros to make them easier to see and
to highlight when a case statement falls though to the next case.

These macros will be important if or when "switch()" is optimized in
the future.