Remove X_PEAS i-code because it offers no saving in code generation, and it makes it harder to write peephole rules that apply to both local and global variables.
Fix instruction rescheduler to stop leaving orphaned I_LDWI behind.

Add I_ISUBWI inverse subtraction for the instruction rescheduler to use.

Rename X_LEA_S to I_LEA_S because it's compiler-generated and not an optimization.

Write the stack-based local variable names into the instruction output as comments
to make debugging easier because the stack offsets are meaningless to humans.

Remove the peephole optimizer's base+offset I_ADDWI rearrangment rules as they are
not used anymore and are just in the way.