Teach HuCC to optimize writing "array[index]" when the array is less than 256 bytes long. Complex index calculations are OK.
This only works for global and static arrays, and not for arrays
that are located on the C stack (-fno-recursive arrays are OK).

Change the "array[variable]" read optimization to handle complex
indexes at the cost of being slightly slower for just "variable".

Reorganize the optimizer's instruction rescheduler to make it a
bit more understandable, and to handle the "array[index]" write.

Move the __stwip and __stbip optimizations after the instruction
rescheduler because they get in its way and the rescheduler gets
better results.