Change how HuCC generates the names for static local variables so that they are symbols and not labels.
Defer the address addition for static local and norecursive arrays so that they
can take advantage of the same peephole optimization rules as global arrays.

Display static and norecursive function local variable names as comments in the
listing file so that it is easier to see what the generated code is doing.

Change how the expression handler decides if an lvalue is unsigned to hopefully
make it a bit less error-prone.

Add some of SmallC-85's comments to expr.c to try to make the source code a bit
more readable.