Add some missing rules for constant expressions to HuCC's peephole optimizer so that they can be recursively calculated and resolved into integers.
Reorganize the optimizer so that math with an integer is optimized early on and
doesn't get in the way of more-complex later optimizations.

Change HuCC's "jsr smul/sdiv" into i-codes to remove the I_JSR and T_LIB string
comparisons that are a PITA to deal with in the optimizer, and add new multiply
and divide by an integer i-codes.

Add some rules to convert signed multiply and divide into unsigned if possible,
and then to convert unsigned power-of-two divide or mod into LSR or AND.

Allow the instruction rescheduler to reorder multiplies as they're commutative.