Change HuCC to split all comparisons and tests into 2 stages, the first to set the flags for branching, and the second to convert the flag into a boolean int that can be stored or used in math.
This allows the second stage I_BOOLEAN i-code to be removed by the peephole if
an integer value is not needed/used.

At the moment this just saves one CLY instruction for most tests, but it could
be more if the test results return the C flag instead of the Z flag as true or
false.