Change the HuCC code generator to split when I_TST_WR and I_BOOLEAN are generated so that I_BOOLEAN doesn't have to preserve the C flag anymore.
This ensures that I_BFALSE and I_BTRUE are always preceeded by an I_TST_WR before
the peephole optimizer does its job, and it means that no workaround is needed if
the optimizer is disabled.

Also add the X_NOT_CF optimization to invert the carry flag for the case if a "!"
follows a conditional enclosed in parentheses.