Add a new I_COMMENT i-code to HuCC that is part of the instruction queue but is ignored by the peephole optimizer.
Change the peephole optimizer to preserve comments when removing instructions.

Change the option that outputs C source into the output file to use this i-code
while a function is being compiled so that peephole optimizations are no longer
spoiled by the queue being flushed when a C source line is output.

Cleanup the function comments in io.c a bit.