Add source-level-debugging information output to PCEAS for the mesen2 emulator.
This is controlled with the "-gC", "-gA", or '-gL' command line flags to choose
whether to output source mappings to the C, ASM, or PCEAS-generated .LST files.

The .LST file gives the most information, because you see the macro expansions,
but the ASM version might be preferable to developers who don't macros much.

The C version will map to C source lines when it can, but there are going to be
mappings to the ASM files for compiler-generated code, and for library code.

Change PCEAS so that equates ('=' and ".equ") can have an optional parameter to
set the label's size (for data), or whether it is "CODE or a "FUNC", which gets
passed to the .sym file for debugging, but is otherwise unused by PCEAS itself.

Change PCEAS to implement the ".DBG line" and ".DBG clear" pseudo-ops to enable
HuCC to tag the original C source file, line and column of the ASM code that it
generates.

Change PCEAS to clean up all of the '\' and '/' characters in file paths so the
paths match the host platform's native format.