Add source-level-debugging information output to HuCC for the mesen2 emulator.
This is controlled with the "-gC", "-gA", or "-gL" command line flags to choose
whether to output source mappings to the C, ASM, or PCEAS-generated .LST files.

The .LST file gives the most information, because you see the macro expansions,
but the ASM version might be preferable to developers who don't macros much.

The C version will map to C source lines when it can, but there are going to be
mappings to the ASM files for compiler-generated code, and for library code.

Change PCEAS to clean up all of the '\' and '/' characters in file paths so the
paths match the host platform's native format.