Change HuCC to use the HuC6280 hardware stack for expression temporaries rather than the zero-page software stack that is used for locals and parameters, since math on the top-of-stack takes the same number of cycles and bytes either way.
This removes the need to dedicate the X register as the software stack pointer,
and so also removes the need for "__xsafe" in __fastcall function declarations.

Remove "__xsafe" from the HuCC compiler now that it is not needed anymore.

Remove the saving of the X register during "#asm" sections.

Simplify the return value from a C function to use the X register.

Change the optimizer now that it does not need to fix software-stack references
when expressions are optimized.

Change all of HuCC's macros and the hucc.asm startup code to work with this new
convention.