Change HuCC's code-generation and optimization to allow rescheduling math with global, stacked and array values as well as the optimization with constants.
Allow the rescheduling of "+=", "-=", "&=", "^=" and "|" to move the access to
the destination variable until the end of the expression, just before the save.

This often removes an I_PUSH_WR and allows for other further optimizations.

Write an I_INFO marker at the beginning of expressions so that whole-expression
optimization can be implemented in the future.